+++ src/editor.h
@@ -41,6 +41,8 @@ extern char current_filename[256];
 extern char *text_buffer[MAX_LINES];
 extern int line_count;
 extern int start_line;
+extern int cursor_x;
+extern int cursor_y;
 extern Node *undo_stack;
 extern Node *redo_stack;
 void handle_regular_mode(int ch, int *cursor_x, int *cursor_y);
@@ -50,7 +52,6 @@ void close_editor();
 void clear_text_buffer();
 void run_editor();
 void initialize_buffer();
-void find(int new_search);
 void redraw(int *cursor_x, int *cursor_y);
 void delete_current_line(int *cursor_y, int *start_line);
 void insert_new_line(int *cursor_x, int *cursor_y, int *start_line);
