+++ src/search.h
@@ -1,0 +1,7 @@
+#ifndef SEARCH_H
+#define SEARCH_H
+
+void find_next_occurrence(const char *word, int *cursor_x, int *cursor_y);
+void find(int new_search);
+
+#endif
